+++ templates/componentes/head.html
@@ -8,7 +8,6 @@
             <li><a href="../favoritos/favoritos.php"><i class="fa fa-heart"></i> <span>1</span></a></li>
             <li><a href="../home/mapa.php"><i class="fa fa-map-marker"></i> </a></li>
             <li><a href="../user/user.php"><i class="fa fa-user"></i></a></li>
-            <li><a href="../admin/home_admin.php"><i class="fa fa-home"></i></a></li>
             <li id="btn-1"><a href="../../inc/peticiones/login/logout.php"><i class="fa fa-sign-out "></i></a></li>   
         </ul>
     </div>
@@ -43,7 +42,7 @@
 <header class="header">
     <div class="container">
         <div class="row">
-            <div class="col-lg-2">
+            <div class="col-lg-3">
                 <div class="header__logo">
                     <a href="../home/home.php"><img src="../../src/img/logo100w57h.png" alt=""></a>
                 </div>
@@ -67,13 +66,12 @@
                         </ul>
                 </nav>
             </div>
-            <div class="col-lg-3">
+            <div class="col-lg-2">
                 <div class="header__cart">
                     <ul>
                         <li><a href="../favoritos/favoritos.php"><i class="fa fa-heart"></i> <span>1</span></a></li>
                         <li><a href="../home/mapa.php"><i class="fa fa-map-marker"></i></a></li>
                         <li><a href="../user/user.php"><i class="fa fa-user"></i></a></li>
-                        <li><a href="../admin/home_admin.php"><i class="fa fa-home"></i></a></li>
                         <li id="btn-2"></li>   
                     </ul>
