+++ templates/componentes/navegacion_reducido.html
@@ -27,17 +27,7 @@
             </div>
             <div class="col-lg-9">
                 <div class="hero__search">
-                    <form id="eleccion_ciudad">
-                        <div class="input-group mb-3">
-                            <select class="custom-select" name="cbx" id="cbx">
-                                <option value="">Selecciona una ciudad</option>
-                            </select>
-                            <div class="input-group-append">
-                                <button id="enviar" name="enviar" value="Guardar" class="btn btn-outline-secondary" type="button">Guardar</button>
-                            </div>
-                        </div>
-                            <!--<input type="submit" id="enviar" name="enviar" value="Guardar" />-->
-                    </form>
+                    
                     <div class="hero__search__form">
                         <form action="#">
                             <input type="text" placeholder="Busca un Restaurante o platillo">
@@ -45,6 +35,17 @@
                         </form>
                     </div>
                 </div>
+                <form id="eleccion_ciudad">
+                    <div class="col-md-5 mt-3 input-group mb-3">
+                        <select class="custom-select" name="cbx" id="cbx">
+                            <option value="">Selecciona una ciudad</option>
+                        </select>
+                        <div class="input-group-append">
+                            <button id="enviar" name="enviar" value="Guardar" class="btn btn-outline-secondary" type="button">Guardar</button>
+                        </div>
+                    </div>
+                        <!--<input type="submit" id="enviar" name="enviar" value="Guardar" />-->
+                </form>
             </div>
         </div>
     </div>
